Use binary-install for npm
