+++ npm/binary.js
@@ -1,0 +1,43 @@
+const { Binary } = require("binary-install");
+const os = require("os");
+const { join } = require("path");
+
+const getPlatform = () => {
+  const type = os.type();
+  const arch = os.arch();
+
+  if (type === "Windows_NT" && arch === "x64") {
+    return "x86_64-pc-windows-msvc";
+  }
+  if (type === "Linux" && arch === "x64") {
+    return "x86_64-unknown-linux-musl";
+  }
+  if (type === "Darwin" && arch === "x64") {
+    return "x86_64-apple-darwin";
+  }
+
+  throw new Error(`Unsupported platform: ${type} ${arch}`);
+};
+
+const getBinary = () => {
+  const platform = getPlatform();
+  const version = require("./package.json").version;
+  const url = `https://workers.cloudflare.com/get-npm-wrangler-binary/${version}/${platform}`;
+  const installDirectory = join(os.homedir(), ".wrangler");
+  return new Binary(url, { name: "wrangler", installDirectory });
+};
+
+const run = () => {
+  const binary = getBinary();
+  binary.run();
+};
+
+const install = () => {
+  const binary = getBinary();
+  binary.install();
+};
+
+module.exports = {
+  install,
+  run
+};
